[graphite] Rasterize SmallPathAtlas paths to the Plot backing store.
Currently in the SmallPathAtlas we rasterize to a separate
SkAutoPixmapStorage and then copy that data to the DrawAtlas via
addToAtlas(). This CL changes that to request space in the DrawAtlas,
and then sets up the SkAutoPixmapStorage to point to the backing
store for the Plot the path is cached in. Finally we use that for our
rasterization step. All of this avoids that unnecessary copy.

Bug: b/294378744
Change-Id: Ic47626da13df8c5ed0f5e1449ae0424620639027
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/820699
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>