Migrate CanvasKit to use modular build
The SkSL tracing code doesn't work because it depends on
SKSL_ENABLE_TRACING being defined or not, which is not compatible
with the modular build. This is not critical to fix for now, but will
need to be done before deleting the GN build.

Change-Id: If564bce32cab9028da9f4cd7aa6d037de8930989
Bug: b/40045064
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/820701
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Owners-Override: Kevin Lubick <kjlubick@google.com>