[graphite] Disable MSAA support on Mac Intel devices.
We are running with MSAA enabled on all devices, which means
that the non-MSAA renderers are not getting tested. Checking
for "Intel" in the name string has been verified on my Mac Mini.

This includes a fix in Device::chooseRenderer(). We were depending
on the internalMultisampleCount ContextOption to decide whether
to use a non-MSAA renderer. However, we also need to make sure that
we can create a MSAA target. In the msaaRequired case
we try to create an MSAA target as well -- checking in
getDefaultMSAATextureInfo() that the target is renderable will
lead us to fall back to single sampling.

Change-Id: Idf7a10a46bed3315cd847e2276554d5f1b9e46b0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/823080
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>