Reland "Fix debugger bazel build"
This reverts commit 3b0b7951146927b3e1d43d753c098f0de2ca48b9.

The original didn't exactly port over the existing behavior of
decoding images into a premul color info. This preserves that
behavior, allowing for the pixel-perfect matching to work in
the test SkMultiPictureDocument_AHardwarebuffer.

The debugger part of CanvasKit still depended on non-modular
builds, and this fixes that.

Change-Id: I55098b54562112028ffa482fb740ad3c79b8a211
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/823656
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>