[graphite] Add Android precompile code path
The new Precompile call is, ultimately, intended to accept an opaque serializable UniqueKey-ish object and create a Pipeline. This CL takes a partial step towards that by working with a GraphicsPipelineDesc and RenderPassDesc - both of which will need to be further broken down (into serializable forms).

Change-Id: I9927c6aa689754c671ca4ff0e37845ef262579e3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/814697
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>