Mark all Test-Fontations jobs as NativeFonts
With recent changes there is now a SkFontMgr_Fontations which can be
used in testing. However, the Fontations jobs are not currently run as
NativeFonts jobs as well, so the new font manager is not yet being used.
Rename the Test-Fontations jobs as Test-NativeFonts_Fontations jobs,
much like the existing DWriteCore jobs.

Bug: b/318667611
Bug: b/328418459
Change-Id: Ifce34de80e57d100b19c686a141bb8348a2df5d7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/822317
Reviewed-by: Dominik Röttsches <drott@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>