Remove SkScalerContextFTUtils::drawCOLRv1Glyph assert
This assert was intended to catch any logic errors when parsing known
good COLRv1 data. Now that this code is tested fairly well this assert
fires more often due to malformed COLRv1 data, but with malformed data
the assert is "expected" to fire. As a result this assert has outlived
its usefulness and is removed.

Change-Id: I9a1a589ebd7687ac8d4f6d7c0a9a92c3857e9fe2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/823658
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>