Change Slug drawing to accept a paint at draw-time
There are situations (linked bug), where the SkPaint can't be fully
realized on the renderer process. To fix that, we allow passing the
SkPaint that was created on the GPU process (via PaintFlags).

For now, there is a single trampoline to maintain the old behavior.
Once chromium is updated, we can stop storing the initial paint in
the slug, entirely.

Bug: b/40071705
Change-Id: I6db67253f94172476f60c0d3873cc1fcfdf74d01
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/824296
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Herb Derby <herb@google.com>