[Fontations-backend] Initial implementation of onGetAdvancedMetrics()
Add ffi for determining whether the font embeddable and subsettable
matching the implementation in the FreeType backend. Add ffi for
retrieving unscaled metrics and add those to the advanced metrics object
that is returned from the new method implementation.

Style information will be added in a separate CL.

With this cahnge
    `$ dm -m 'SkPDF_FontCanEmbedTypeface' --fontations`
passes.

Bug: skia:329377305
Change-Id: I5ddd14ba8d79a82d265fc74845595e72566e289f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/825556
Reviewed-by: Dominik Röttsches <drott@google.com>
Commit-Queue: Dominik Röttsches <drott@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>