[Fontations-backend] Move typeface initialization to Make factory
Move the initialization of the `SkTypeface_Fontations` to the
`MakeFromData` call, return nullptr if creation of the Rust font
reference fails. When construction suceeds, create other required
members and pass everything to the constructor.

Follows the approach in the DirectWrite and FreeType backends.

Preparation for landing a fix for issue 329376987, Fontations typeface
style awareness.

Bug: skia:329439348
Change-Id: Ie28a4458e8d95c0ff16b6e9e8d5fcf255d8c6d71
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/826556
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: Dominik Röttsches <drott@google.com>
Commit-Queue: Dominik Röttsches <drott@google.com>