[gm] Add another stroked colinear cubic case
Skia's rendering of a stroked cubic with a very short near-cusp segment
is inconsistent between the vicinity of a 180-degree turn vs larger
angles. At-or-near-180 turns are special cased by Skia's built-in
stroker but contain a semi-circular hole at smaller angles when rendered
by the raster backend. Vello's behavior currently matches that of the
raster backend but also in the at-or-near-180 case.

Added a new stroke to the trickycubicstrokes GM that demonstrates this.

Change-Id: I8ed70b55742fba2b041c8e7f9ef01b530da2a9f8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/828151
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>