[graphite] Add a RawImage entry point to the Precompilation system
This is mainly for parallelism with the normal Skia API.

Note that expanded testing (i.e., the addition of random_local_matrix) meant that a LocalMatrixShader option needed to be added to the
ImageShader precompilation path.

This CL also takes care of a lingering TODO re adding Image and RawImage entry points to the public SkShaders API. I'm not married to the idea so lmk wdyt.

Bug: b/259548724
Bug: b/40044529
Change-Id: I3092f85814b73b05eeff29c4d36f9dcb451a774c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/826796
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>