[graphite] Propagate isConstant through LocalMatrix and CTM Shaders
The Constantness of the SkColorShader and SkColor4Shader isn't changed by being wrapped in a matrix-changing shader so propagate the property through (just as we do for asGradient).

This came up bc the Precompilation system elides LocalMatrixShaders less often than the normal API. This was resulting in program differences when the normal API elided a wrapping shader. In particular, the ColorShaders' Constantness is solely used to skip dithering.

Bug: b/259548724
Change-Id: I21e88ce0817e8adea8c69e4dd65b0d9467a885ff
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/829176
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>