[graphite] TaskList uses TArray, no more prepending
Removes use of std::vector in TaskList in favor of faster TArray.
TArray doesn't support insertion, so this reorganizes how the tasks
are collected by the draw buffer manager.

With the changes to Recorder's snap(), the overall task order should
be about the same as before: buffer manager uploads, followed by all
the other tasks that depend on the buffers.

The main difference is that because the old code used prepend() for
the DBM tasks, those were actually executed in reverse order from
when they were first created/filled out in the DBM. They should all
be independent so it doesn't really matter.

The underlying allocation pattern should be better now, too, since
we aren't having to do an O(N) shift for every DBM task, and while
it had been a free move initially for the root task list into the
Recording, the new `move_back` should work out to a pure memcpy.
That said, it's unlikely that the task counts are so high that this
ever mattered.

Change-Id: Id42d0c69f18a3c5f10a3fe3d8aad1bd4e6381c2d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/829559
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>