[graphite] Free mapped buffers in submit() and performDeferredWork()
Contrary to ganesh, graphite does not poll in Context::submit() or
performDeferredWork(). The consequence is that mapped buffers are not
freed, increasing memory usage.

In Chromium, this can cost a lot of memory as it performs CPU readback
for pages with (HTML5) <canvas> elements, and the buffer used stays
around.

This CL aligns Graphite's behavior on Ganesh's, and solves the issue
seen in Chromium.

Fix found by sunnyps@ and egdaniel@

Bug: chromium:330504504
Change-Id: Iec2efb27dbbdf3b778e8243dde53d34e654df13e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/830997
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Benoit Lize <lizeb@google.com>