SkPath: reduces cost of addRect
While looking over this code to understand a regression
I figured it's a good time to improve the performance of
it.

This makes addRect() create the editor and update
the PathRef directly rather than calling 4
functions that create editors and mutate the pathref.

This also removes some outdated comments.

Bug: None
Change-Id: I2270cf13ff6d71ac969e22e14547c4a2a8541904
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/832336
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>