Break up Bidi based on ICU or ICU subset into classes to avoid #define
SkUnicode_icu_bidi.cpp had some logic #ifdef'd out related to if there
was a full version of ICU or not. The subset of ICU is needed to do
some bidi calculations for the libgrapheme and "client ICU" versions
of SkUnicode.

There is a slight change now, where we will *always* compile in the
ICU subset (which has a _skia suffix on the functions to avoid ODR issues)
as opposed to before where if a client had both "full ICU" and libgrapheme,
they wouldn't need to compile in the subset. This seems unlikely to
affect anything in practice since clients will probably only want
one implementation of SkUnicode anyway.

Bug: b/40045064
Change-Id: I9b1b5d9a7804e53983c0b9915ee4edaa5cdb14ee
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/831801
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>