Remove unhelpful assert in backend caps format tables.
These checks just force every format that is compatible with a
colorType to be given as an option for the default format for a
colorType. This isn't actually something we want to enforce because
there are some formats (e.g. compressed formats) that we don't ever
want to be a default option.

Bug: b/267626406
Change-Id: Iafd8b5bb149af426428a958a2f880a235349c769
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/833417
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>
Auto-Submit: Greg Daniel <egdaniel@google.com>