Fix defines for icu_subset
Disabling the version suffix, but not supplying a lib suffix
caused the ICU headers to try to re-define the "icu" namespace.

This reworks the skunicode BUILD.gn rules to isolate the ICU
specific defines that were causing issues. We now have several
builds of skunicode that get bundled together (much like how
the Bazel build works).

This also fixes the plain text editor application, which was
not using ICU/Harfbuzz as expected.

Change-Id: Ia15dc9ad5d1da826608e6b4f9a137009fb5ed6bc
Bug: b/40045064
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/833798
Reviewed-by: Ben Wagner <bungeman@google.com>