[graphite] Track linked device on Images
Image views of Surfaces now automatically flush if they are drawn and
the surface has pending work.

Graphite-internal Surfaces don't do this pattern, they still rely on
the automatic flush when the surface goes out of scope. In theory, we
could move to not doing this but that would possibly defer the
work until a later recording (if that's the first time the image view
was drawn). If the surface wraps a backend texture and goes out of
scope, the client is relying on the work being flushed then as well,
so for now I think the flush-on-Device deletion should remain.

In subsequent CLs, special images will wrap a regular graphite-backed
Image so that uninstantiated Devices can also have their tasks tracked.

Bug: b/323887207
Change-Id: I60d5cedd602669ab8266155ee3fe0db8ee80da64
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/833437
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>