Add VkComponentMapping to VulkanYcbcrConversionInfo
VK_FORMAT_G8_B8_R8_3PLANE_420_UNORM is YCbCr and
AHARDWAREBUFFER_FORMAT_YV12 is YCrCb.  We need to swap R and B when
AHARDWAREBUFFER_FORMAT_YV12 is not specified as an external format but
as VK_FORMAT_G8_B8_R8_3PLANE_420_UNORM, which can happen when Android
runs inside a VM and skia is used in the host.

Bug: b/329151784
Change-Id: Ia33e79470cf38a3cf6c8decfc774f49ef6402673
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/836758
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Chia-I Wu <olv@google.com>