[graphite] Relax assert in ~Device
This was overly strict for scratch devices that encountered error
conditions. Confirmed that this resolves the asserts affecting the
Android protected job (which happened to be the source of errors due
to mismatch protected statues). Reproduced by adding the equivalent
protected validation to MtlTexture and marking MtlCaps as requiring
protected content.

Bug: b/323887207
Change-Id: I2b14d7a097e7663c64dd1eab2086034616a083a9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/836557
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>