Fix linking of Frameworks in Mac Objective C builds with Bazel
We were seeing errors like
    ld64.lld: error: framework not found for -framework MetalKit
with no hints from the linker as to why. (Couldn't figure out how
to have the linker tell us where it was looking for the frameworks).

As it turns out, we were not setting the -F flag [1] when linking,
only compiling. I think we had tried it before to no success, however
I learned that the -F flag's path when linking is relative to
the syslibroot (we might have been using the full path as when
compiling)

Once -F/Frameworks was set, I saw new errors along the lines of
    error: unable to locate re-export with install name /System/Library/Frameworks/CoreFoundation.framework/Versions/A/CoreFoundation

Re-export paths are hard-coded based on the original install path.
To address this, I changed where we were symlinking the Frameworks
to - from /symlinks/xcode/MacSDK/Frameworks to
/symlinks/xcode/MacSDK/System/Library/Frameworks. This emulates
the typical install path, which made the re-exports work at link
time because the hard-coded paths now match the path in the syslibroot.

Finally, I was seeing errors like:
    ld64.lld: error: undefined symbol: typeinfo for GrAttachment
    >>> referenced by [...]libganesh_metal.a(GrMtlAttachment.o):(symbol typeinfo for GrMtlAttachment+0x10)
I saw these for any subclass we had made in .mm files when the super
class was implemented in .cpp files. This was due to RTTI being
turned off, so I turned it back on for Mac and iOS builds.

[1] "Add directory to framework include search path"

Change-Id: I894b61dd2e8d857ddbf9277c25d0ad99824260aa
Bug: b/256860816
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/836797
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>