Ensure inliner adds braces around scopes where required.
Previously, if both the if-condition and the if-body statement
were inlined, and the body statement lacked braces, the inliner
would neglect to add braces around the body. This would lead to
a multi-statement if-body without braces, which would either
assert or generate bad code.

Change-Id: I6efe25bef6a41fd219db77c70242308a9356e9b5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/837556
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>