[graphite] Fix subsetting for UV planes.
When the Y plane has odd dimensions and the subsampling values for U
and V are 2 or 4, we end up with UV textures that have some extra
texels on the right and bottom edges. The coords that we send for
sampling are in Y space and we'll end up dividing by Y's dimensions,
giving us a range of [0,1], which will end up oversampling those extra
pixels in U and V. The solution is to send up an extra set of
dimensions which we'll use to transform the unnormalized texture
coordinates in Y space into normalized coordinates in UV's texel space.

Bug: b/283271538
Change-Id: I1b9aba27abc60dbe0bb0167db6c761cbedaafdad
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/836936
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>