Fix shaping in GN builds
As a follow-up to https://skia-review.googlesource.com/c/skia/+/826067
this fixes the DEPs of DM, nanobench, and viewer such that they
get the defines set by skunicode.

Note that in GN, "defines" only flow to dependencies
(e.g. fromskunicode -> DM) if there is a direct dependency or an
intermediate dependency has as a public_dep on skunicode
(e.g. skunicode -> skshaper -> DM). [1]
This CL establishes the former rather than the latter because it is
more explicit and closer to how Bazel works (and I wasn't sure of
the blast radius of changing skshaper's dep to a public_dep).

[1] https://gn.googlesource.com/gn/+/main/docs/reference.md#var_defines

Change-Id: Icb5aa39653c801ed79be00bb99c54a5cb8a202fa
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/837516
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>