Remove GLSL/Metal/WGSL indentation in release mode.
Indentation only exists for the benefit of human readers.
In skslc standalone mode, or in debug, it is worth emitting the
indentation. In release mode, adding a bunch of spaces to the
front of each line is throwing away memory and CPU cycles.

As a data point, I measured the size of the finished WGSL vertex
shader of `addarc` in Viewer. The generated code was 16360 bytes
originally, and 13080 bytes without leading indentation.
Shockingly, 20% of the bytes in the finished shader consisted of
leading spaces! (See: http://gpaste/5223517091528704 . This is
just Graphite's `tessellate_stroked_curve` after inlining.)

This will affect how shader compilation errors look if they occur
in the wild. We could apply SkShaderUtils::PrettyPrint to the
output shader, but that would invalidate any line numbers in the
source, so it's safer to leave it as-is and reformat them manually
when needed.

Change-Id: Ia385ac39b84b93a509c620618ef0b7145e6c0d96
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/837598
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>