[pdf] Allow OpenType CFF fonts to be embedded
Sfntly did not subset CFF tables and so it was generally better to embed
them as Type3 fonts instead of embedding the whole unsubset file. Now
that the HarfBuzz subsetter is being used and can subset the CFF table,
allow CFF flavor OpenType fonts to be embedded.

Change-Id: I3fd3abf07c5e860a7aa95cbeb800a0fad3406594
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/837698
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>