[graphite] Add linear filter inset for YUV.
Consider a logical image pixel at the edge of a subset. When computing
the logical pixel color value we should use a blend of two values
from the subsampled plane. Depending on where the subset edge falls in
actual subsampled plane, one of those values may come from outside the
subset. Hence, we use this custom inset which applies the wrap mode to
the subset but allows linear filtering to read pixels from the plane
that are just outside the subset.

Bug: b/283271538
Change-Id: Ieb87453f3e8171dc58de4f108059d71fbfccc19a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/838099
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>