[graphite] Expand DrawTypeFlags to allow finer control of options
For ImageFilters in particular we know a lot about how we're going to draw the intermediate steps and the final result. Being able to limit the renderSteps that are precompiled for ImageFilters can greatly reduce the combinatorial explosion.

This CL just adds the new types but doesn't broadly use them. That is reserved for the upcoming blur ImageFiltering CL.

Bug: b/259548724
Change-Id: I192d79f9c2a1c979533e79647715ff695431a0ee
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/838636
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>