Properly convert between axis definition types
The AxisDefinition fMinimum, fDefault, fMaximum were changed from
SkFixed to SkScaler. However, the code using these values was not
updated to convert between these as expected. Technically this worked
out due to a number of automatic conversions and the old scaling code
left in place, but leads to great difficulty in understanding the code.

Change-Id: I3597a8414578c072716ebc38646ad27967a82972
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/838736
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>