[Graphite] Reduce calls to processReturnedResources in ResourceCache.
processReturnedResources can have measureable overhead each time it
grabs and releases the return mutex. So reducing the amount of times we
call can help some performance. First, this change causes it so that we
don't return resources when inserting 0 sized resources. Second, when
finding a new resource we first search for an available resources before
trying to return resources to the cache.

Change-Id: I68e03af04e007b09a9662b36c866066112b0fe00
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/838679
Commit-Queue: Greg Daniel <egdaniel@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>