Add job to test with checked-in FreeType
Currently all the Linux based test jobs are using the system FreeType.
The Android_NativeFonts job is currently the only job testing this
configuration, but as it is GPU only is not running all the tests. Add a
job to test on Linux with the newer checked-in version of FreeType.

Change-Id: Ib23f909ad99c1cf4742bd6ee807518421ff44f7c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/839271
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>