Reland "Migrate SkParagraph's tests etc to provide SkUnicode"
This is a reland of commit 6184d91ed4e9c641fdfbf6ea1cf23c533e2fc252

To handle the client icu implementation (in which for every call
to make a ParagraphBuilderImpl, we must make a new "SkUnicode" instance,
which results in calling out to the browser's APIs), CanvasKit
itself needs to pass in the SkUnicodes::Client::Make callback
instead of it being in an #ifdef in ParagraphBuilderImpl.cpp

This CL feeds that callback through with a backwards compatible
implementation until other clients (e.g. skwasm) are updated.
jlavrova has promised to clean up the other #ifdefs related to
the client ICU impl in a few weeks. Until that happens, only
the GN build of CanvasKit will work properly with the client ICU
data (Bazel will not). This should be ok, given that our tools
(e.g. shaders.skia.org) are perfectly fine with the full ICU
implementation.

Original change's description:
> Migrate SkParagraph's tests etc to provide SkUnicode
>
> It was convenient to use the SkShaper::Factory stuff landed in
> https://skia-review.googlesource.com/c/skia/+/826067 to avoid
> copying more #ifdefs around.
>
> Bug: b/40045064
> Change-Id: I17b0d3679c15d3c401fbf08106123b52e3cbdee1
> Reviewed-on: https://skia-review.googlesource.com/c/skia/+/837700
> Commit-Queue: Kevin Lubick <kjlubick@google.com>
> Reviewed-by: Julia Lavrova <jlavrova@google.com>

Bug: b/40045064
Change-Id: I5e4cafb601d7d3603277c7497817450b7ac1afae
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/838417
Reviewed-by: Ben Wagner <bungeman@google.com>