Add isArc to SkPath
This remembers when a path is a simple elliptical arc (ie, it could be
rendered by a call to SkCanvas::drawArc).

For now, this only remembers perimeter arcs (when useCenter is false).
I intend to support useCenter arcs as well, but those will be more
complicated. Regardless, the new interface will support both uses.

Bug: b/330506337
Change-Id: I5c43809e0d57d2e1a4e5704d4027863412d6a739
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/841396
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>