Fail SkParagraph GMs gracefully when unicode fails to be created
This was causing an issue on our OptimizeForSize performance
benchmarks. I tried to use the client ICU build, but that didn't
work, probably because it's meant to be browser-only. So I just
had the gm skip doing anything with the paragraph in that case.

Change-Id: Id11f3946205b7d53122b81053b25a7dccc0c7170
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/840857
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>