[graphite] Add minimal API to specify precompilation ImageFilters
This minimal API is just being added to support the changes to PaintOptions::buildCombinations. That method now creates the pipelines for the blur image filter. To accomplish this the PaintOptions::ProcessCombination function was expanded to take new parameters that the blur code wants to override (i.e., drawTypes, primitiveBlender and coverage).

As part of this I've also updated the code to be more precise about which variables are part of the lamda(s).

Bug: b/259548724
Change-Id: If097f8d03461645af514bbe3199873e86cbae52c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/840858
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>