Allow old CoreText to return wrong style
In 10.15 and earlier, CoreText applied the value of the `wdth` axis to
the computed style differently. Until 10.15 the value of the `wdth` axis
was the old GX definition. With macOS 11 it is now interpreted with the
new OpenType definition. Allow these old versions of macOS to continue
to report their old values.

Change-Id: I6db765445497a9da4b5633b4196c185f90c66a81
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/841656
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>