Move to vpython 3.11
Previously with "Roll depot_tools, recipes-py, python 3.11" an attempt
was made to roll everything to a newer version to use 3.11 exclusively.
Something about that change caused issues with starting jobs, so instead
roll each of these separately to narrow down the cause.

With this change the Skia specific parts of the build will run with
3.11, but recipies will still run with 3.8.

Change-Id: I9de37d13d8e5c147d1703004863964bc0e2b4802
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/842480
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>