[graphite] Don't store unmappable buffer as reusable
As the original logic was written, the first time through the gate, the
bad mapping was checked and null would be returned. But `fReusedBuffer`
still pointed to the buffer that failed to map its pointer and so on
subsequent "reuses", we'd either hit the assert or assume it was valid
and add an offset that then tripped up BufferWriter's null pointer
checks.

Adds validate() guards to writeRGBFromRGBx() and updates the
validation checks to take into account the actual offset.

Bug: b/336099925
Bug: b/336245801
Change-Id: I9dd0cf72a4b452248fe45120b57f72c0e2dc8a65
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/844316
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>