Do not create the image filter cache just to purge it
Creation of the cache is deferred until it's used, but today, that
includes purging the cache. This means that if we haven't used it at
all, we create it just to purge it.

Add a parameter to Get to determine whether it should be created if it
doesn't already exist. When purging, do not create it.

Bug: 327992219
Change-Id: I8c8822b726ab3db989dcb5dd48ec6d3c4d97c649
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/844796
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>